[Back]_Feat_FIx_Rename_Update 백엔드 이슈응답의 카운팅 부분 작업, 용어 변경에 따른 파일 이동및 수정, stats 영역 코드정리
